+++ stroom-config/stroom-config-global-impl/src/test/java/stroom/config/global/impl/GenerateConfigProvidersModule.java
@@ -7,7 +7,11 @@ import stroom.proxy.repo.RepoDbConfig;
 import stroom.util.logging.LambdaLogger;
 import stroom.util.logging.LambdaLoggerFactory;
 import stroom.util.shared.AbstractConfig;
-import stroom.util.shared.BootStrapConfig;
+import stroom.util.shared.IsProxyConfig;
+import stroom.util.shared.NotInjectableConfig;
+
+import com.google.common.reflect.ClassPath;
+import com.google.common.reflect.ClassPath.ClassInfo;
 
 import java.io.File;
 import java.io.IOException;
@@ -21,6 +25,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
 /**
@@ -30,6 +35,8 @@ import java.util.stream.Collectors;
 public class GenerateConfigProvidersModule {
 
     private static final LambdaLogger LOGGER = LambdaLoggerFactory.getLogger(GenerateConfigProvidersModule.class);
+
+    static final String THROWING_METHOD_SUFFIX = "ButThrow";
 
     private static final String CLASS_HEADER = """
             package stroom.config.global.impl;
@@ -54,7 +61,7 @@ public class GenerateConfigProvidersModule {
                 // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
             """;
 
-    public static void main(String[] args) {
+    public static void main(String[] args) throws IOException {
         final ConfigMapper configMapper = new ConfigMapper();
         final Set<String> simpleNames = new HashSet<>();
         final Map<String, List<String>> simpleNameToFullNamesMap = new HashMap<>();
@@ -67,10 +74,30 @@ public class GenerateConfigProvidersModule {
         final String methodsStr = configMapper.getInjectableConfigClasses()
                 .stream()
                 .sorted(Comparator.comparing(Class::getName))
-                .filter(clazz ->
-                        !clazz.isAnnotationPresent(BootStrapConfig.class))
                 .map(clazz ->
                         buildMethod(simpleNames, simpleNameToFullNamesMap, clazz))
+                .collect(Collectors.joining("\n"));
+
+        final Predicate<String> packageNameFilter = name ->
+                name.startsWith("stroom") && !name.contains("shaded");
+
+        // Methods that throw for non injectable config
+        // We don't want the risk of a dev accidentally injecting a NotInjectable config
+        // with the resulting confusion it will cause when it has the wrong values so make
+        // providers that always throw.
+        // TODO 16/12/2021 AT: Change to use IsStroomConfig in >=7.1
+        final ClassLoader classLoader = GenerateConfigProvidersModule.class.getClassLoader();
+        final String notInjectableMethodsStr = ClassPath.from(classLoader)
+                .getAllClasses()
+                .stream()
+                .filter(classInfo -> packageNameFilter.test(classInfo.getPackageName()))
+                .map(ClassInfo::load)
+                .filter(clazz -> clazz.isAnnotationPresent(NotInjectableConfig.class))
+                .filter(AbstractConfig.class::isAssignableFrom)
+                .filter(clazz -> !IsProxyConfig.class.isAssignableFrom(clazz))
+                .map(clazz ->
+                        buildThrowingMethod(simpleNames, simpleNameToFullNamesMap,
+                                (Class<? extends AbstractConfig>) clazz))
                 .collect(Collectors.joining("\n"));
 
         final String repoConfigMethodStr = buildMethod(
@@ -91,6 +118,10 @@ public class GenerateConfigProvidersModule {
                 repoDbConfigMethodStr,
                 SEPARATOR,
                 methodsStr,
+                """
+                            // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+                        """,
+                notInjectableMethodsStr,
                 CLASS_FOOTER);
 
         updateFile(fileContent);
@@ -145,6 +176,52 @@ public class GenerateConfigProvidersModule {
                 fullReturnClassName,
                 methodNameSuffix,
                 fullInstanceClassName);
+    }
+
+    private static String buildThrowingMethod(final Set<String> simpleNames,
+                                              final Map<String, List<String>> simpleNameToFullNamesMap,
+                                              final Class<? extends AbstractConfig> clazz) {
+        final String simpleClassName = clazz.getSimpleName();
+        // Fix the name for nested classes
+        final String fullClassName = clazz.getName()
+                .replace("$", ".");
+
+        simpleNameToFullNamesMap.computeIfAbsent(simpleClassName, k -> new ArrayList<>())
+                .add(fullClassName);
+
+        String methodNameSuffix = simpleClassName;
+        int i = 1;
+        while (simpleNames.contains(methodNameSuffix)) {
+            methodNameSuffix = simpleClassName + ++i;
+        }
+        if (i != 1) {
+            LOGGER.warn("Simple class name {} is used by multiple classes {}, using method name suffix {}",
+                    simpleClassName,
+                    simpleNameToFullNamesMap.get(simpleClassName),
+                    methodNameSuffix);
+        }
+        simpleNames.add(methodNameSuffix);
+
+        final String template = """
+                    @Generated("%s")
+                    @Provides
+                    @SuppressWarnings("unused")
+                    %s get%s%s(
+                            final ConfigMapper configMapper) {
+                        throw new UnsupportedOperationException(
+                                "%s cannot be injected directly. "
+                                        + "Inject a config class that uses it or one of its sub-class instead.");
+                    }
+                """;
+
+        return String.format(
+                template,
+                GenerateConfigProvidersModule.class.getName(),
+                fullClassName,
+                methodNameSuffix,
+                THROWING_METHOD_SUFFIX,
+                fullClassName);
+
     }
 
     private static void updateFile(final String content) {