+++ stroom-query/stroom-query-common/src/main/java/stroom/query/common/v2/Store.java
@@ -17,6 +17,7 @@
 package stroom.query.common.v2;
 
 import java.util.List;
+import java.util.Set;
 import java.util.concurrent.TimeUnit;
 
 public interface Store {