+++ stroom-search/stroom-search-extraction/src/main/java/stroom/search/extraction/StreamEventMap.java
@@ -4,12 +4,10 @@ import stroom.util.concurrent.CompleteException;
 import stroom.util.logging.LambdaLogger;
 import stroom.util.logging.LambdaLoggerFactory;
 
-import java.util.AbstractMap.SimpleEntry;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.Map;
-import java.util.Map.Entry;
 import java.util.Set;
 import java.util.concurrent.locks.Condition;
 import java.util.concurrent.locks.Lock;
@@ -19,14 +17,18 @@ public class StreamEventMap {
 
     private static final LambdaLogger LOGGER = LambdaLoggerFactory.getLogger(StreamEventMap.class);
 
+    private static final Key COMPLETE = new Key(-1, -1);
+
     private final Map<Long, Set<Event>> storedDataMap;
-    private final LinkedList<Long> streamIdQueue;
+    private final LinkedList<Key> streamIdQueue;
     private final int capacity;
     private int count;
 
     private final Lock lock = new ReentrantLock();
     private final Condition notFull = lock.newCondition();
     private final Condition notEmpty = lock.newCondition();
+
+    private long extractionDelayMs;
 
     public StreamEventMap(final int capacity) {
         this.storedDataMap = new HashMap<>();
@@ -40,8 +42,9 @@ public class StreamEventMap {
             while (count == capacity) {
                 notFull.await();
             }
-            streamIdQueue.addLast(-1L);
-            ++count;
+
+            streamIdQueue.addLast(COMPLETE);
+            count++;
             notEmpty.signal();
         } finally {
             lock.unlock();
@@ -54,17 +57,18 @@ public class StreamEventMap {
             while (count == capacity) {
                 notFull.await();
             }
-            final Set<Event> eventSet = storedDataMap.compute(event.getStreamId(), (k, v) -> {
+
+            final Set<Event> events = storedDataMap.compute(event.getStreamId(), (k, v) -> {
                 if (v == null) {
                     // The value is null so this is a new entry in the map.
                     // Remember this fact for use after we have added the new value.
                     v = new HashSet<>();
-                    streamIdQueue.addLast(k);
+                    streamIdQueue.addLast(new Key(k, System.currentTimeMillis()));
                 }
                 return v;
             });
-            if (eventSet.add(event)) {
-                ++count;
+            if (events.add(event)) {
+                count++;
                 notEmpty.signal();
             } else {
                 LOGGER.warn("Duplicate segment for streamId=" +
@@ -77,35 +81,96 @@ public class StreamEventMap {
         }
     }
 
-    public Entry<Long, Set<Event>> take() throws InterruptedException, CompleteException {
-        Entry<Long, Set<Event>> entry = null;
+    public EventSet take() throws InterruptedException, CompleteException {
+        Key key;
+        EventSet eventSet = null;
+        long delay = 0;
+
         lock.lockInterruptibly();
         try {
             while (count == 0) {
                 notEmpty.await();
             }
 
-            final long streamId = streamIdQueue.removeFirst();
-            if (streamId == -1) {
-                streamIdQueue.addLast(-1L);
-                notEmpty.signal();
-            } else {
-                final Set<Event> events = storedDataMap.remove(streamId);
-                entry = new SimpleEntry<>(streamId, events);
-                count -= events.size();
+            key = streamIdQueue.peekFirst();
+            if (key != null) {
+                if (key == COMPLETE) {
+                    notEmpty.signal();
+                } else {
+                    delay = extractionDelayMs - (System.currentTimeMillis() - key.createTimeMs);
+                    if (delay <= 0) {
+                        key = streamIdQueue.removeFirst();
+                        final Set<Event> events = storedDataMap.remove(key.streamId);
+                        eventSet = new EventSet(key.streamId, events);
+                        count -= events.size();
+                        notFull.signal();
+                    }
+                }
             }
-            notFull.signal();
         } finally {
             lock.unlock();
         }
 
-        if (entry == null) {
+        if (key == COMPLETE) {
             throw new CompleteException();
         }
-        return entry;
+
+        if (delay > 0) {
+            Thread.sleep(delay);
+        }
+
+        return eventSet;
     }
 
     public int size() {
         return count;
+    }
+
+    public void setExtractionDelayMs(final long extractionDelayMs) {
+        this.extractionDelayMs = extractionDelayMs;
+    }
+
+    private static class Key {
+
+        private final long streamId;
+        private final long createTimeMs;
+
+        public Key(final long streamId,
+                   final long createTimeMs) {
+            this.streamId = streamId;
+            this.createTimeMs = createTimeMs;
+        }
+
+        public long getStreamId() {
+            return streamId;
+        }
+
+        public long getCreateTimeMs() {
+            return createTimeMs;
+        }
+    }
+
+    public static class EventSet {
+
+        private final long streamId;
+        private final Set<Event> events;
+
+        public EventSet(final long streamId,
+                        final Set<Event> events) {
+            this.streamId = streamId;
+            this.events = events;
+        }
+
+        public long getStreamId() {
+            return streamId;
+        }
+
+        public Set<Event> getEvents() {
+            return events;
+        }
+
+        public int size() {
+            return events.size();
+        }
     }
 }