+++ stroom-search/stroom-search-impl/src/main/java/stroom/search/impl/SearchConfig.java
@@ -13,7 +13,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyDescription;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
-
 @JsonPropertyOrder(alphabetic = true)
 public class SearchConfig extends AbstractConfig implements IsStroomConfig {